[Qt] Introduce new qmake variable 'WEBKIT' for signaling dependencies
The custom qmake variable 'WEBKIT' is used for signaling that a
target depends in some way on other subproject of the WebKit
project. For now this is limited to the set of intermediate
libraries: wtf, javascriptcore, webcore, and webkit2.

This replaces the previous convension of using load(foo) for
just include paths, and CONFIG += foo to also link against foo.

Adding a dependency results in additional include paths being
available, and potentially linking to the library. This is
decided by the build system based on conditions such as what
kind of target is being built and the general build config.

An advantage to his approach is that it simplifies the individual
foo.prf files, for example by allowing us to use INCLUDEPATH +=
and LIBS += as normal instead of prepending.

Reviewed by Simon Hausmann.

Identifier: 92321@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104034 268f45cc-cd09-0410-ab3c-d52691b4dbfc