REGRESSION (r90797): Full screen video HUD cannot be dragged horizontally https://bugs.webkit.org/show_bug.cgi?id=75200
Reviewed by Eric Carlson.

Source/WebCore:

Test: fullscreen/video-controls-drag.html

The !important rules in fullscreenQuickTime.css are overriding the styles added by the
drag operation in MediaControlElements.cpp.  Give the panel a "dragged" class in setPosition
(clearing it in resetPosition) that allows the !important rules to apply only when the
panel is not dragged.

* css/fullscreenQuickTime.css:
(video:-webkit-full-screen::-webkit-media-controls-panel):
(video:-webkit-full-screen::-webkit-media-controls-panel:not(.dragged)):
* html/shadow/MediaControlElements.cpp:
(WebCore::MediaControlPanelElement::setPosition):
(WebCore::MediaControlPanelElement::resetPosition):

LayoutTests:

* fullscreen/video-controls-drag-expected.txt: Added.
* fullscreen/video-controls-drag.html: Added.

Identifier: 92511@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104301 268f45cc-cd09-0410-ab3c-d52691b4dbfc