Make it possible to type data: URLs into MiniBrowser on Windows
Fixes <http://webkit.org/b/75084> Crash when trying to navigate to a data: URL in
MiniBrowser on Windows

Reviewed by Darin Adler.

* MiniBrowser/win/BrowserWindow.cpp:
(BrowserWindow::handleMessage): Instead of checking for an "http://" prefix when deciding
whether to prepend "http://", just look for whether the typed URL contains a colon. If it
does, we assume it already has a scheme and don't modify it.


Identifier: 92798@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104621 268f45cc-cd09-0410-ab3c-d52691b4dbfc