[Qt] Change how build-webkit decides when to do full incremental builds
Instead of relying on update-webkit (which isn't run on the bots) to
decide when to do a full incremental build (make qmake), we let the
build-webkit script itself check the current SVN revision against the
previous build (by storing it in .webkit.config).

If the two differ we assume a full incremental build is needed, since
the new revisions might have introduced problematic things like new
Q_OBJECT macros. If not, we assume the developer is doing changes
locally, and revert to doing a plain 'make'.

In addition, when the build fails in the latter case, we inform the
developer of possible pitfalls and how to manually run 'make qmake'.

Reviewed by Ossy.

Identifier: 93637@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@105603 268f45cc-cd09-0410-ab3c-d52691b4dbfc